added number #

+++ src/hometasks/Fibonacci/FibNumbers.java
@@ -7,7 +7,7 @@ public class FibNumbers {
 
 	public static void main(String[] args) {
 		System.out.println("Fibonacci numbers:");
-		BlockingQueue<Integer> queue = new ArrayBlockingQueue<>(FibNumbersGenerator.ITERATIONS);
+		BlockingQueue<Long> queue = new ArrayBlockingQueue<>(FibNumbersGenerator.ITERATIONS);
 		Thread generator = new Thread(new FibNumbersGenerator(queue));
 		Thread printer = new Thread(new FibNumbersPrinter(queue));
 		generator.start();
